+++ yike-20260707/alibabacloud_yike20260707/models/_list_asset_categories_response_body.py
@@ -12,7 +12,7 @@ class ListAssetCategoriesResponseBody(DaraModel):
         self,
         categories: List[main_models.ListAssetCategoriesResponseBodyCategories] = None,
         request_id: str = None,
-        total: str = None,
+        total: int = None,
     ):
         self.categories = categories
         self.request_id = request_id